Specify global non POD g_argvs with static to fix ODR violation
This was exposed by ASAN when dynamically linking libgtest.so and
libgmock.so (build without gtest_hide_internal_symbols cmake flag, which
is off by default) with a test program and running the program. At the
end of the test execution, both gtest and gmock tries to destroy the
global variable `g_argvs`. Specify it with `static` to fix the problem.